CTX-4912: Fixed typo in directoy name [documents -> document]
