+++ document-ocr-fn/resources/function/detect_document.py
