+++ document-ocr-fn/resources/function/image_segmentation.py
