+++ document-ocr-fn/resources/function/model.py
